test(stat): recategorize and clean up browser tests
